refactor: move the infoIconPopup out of the temp folder
