+++ src/features/projectsV2/ProjectDetails/stories/InfoIconPopup.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import InfoIconPopup from '../../components/InfoIconPopup';
+import InfoIconPopup from '../components/microComponents/InfoIconPopup';
 
 const meta: Meta<typeof InfoIconPopup> = {
   title: 'Projects/Common/InfoIconPopup',
